+++ packages/plugin-types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/plugin-types
 
+## 1.2.1
+
+### Patch Changes
+
+- a65371c: upgrade aws-cdk and aws-cdk-lib to ^2.152.0
+
+## 1.2.0
+
+### Minor Changes
+
+- d9b83a1: support adding more than one bucket
+
+## 1.1.1
+
+### Patch Changes
+
+- 3c698e0: upgrade AWS SDK packages to latest
+
 ## 1.1.0
 
 ### Minor Changes